+++ sale_stock_prebook_stock_available_to_promise_release/models/__init__.py
@@ -1,2 +1,3 @@
 from . import stock_move
-from . import sale
+from . import sale_order
+from . import sale_order_line
